user_team_service/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

psycopg2-binary==2.9.9

pytest==8.0.2
pytest-django==4.7.0
pytest-cov==4.1.0

drf-yasg==1.21.7

celery==5.3.6
redis==5.0.1

python-dotenv==1.0.0
django-environ==0.11.2